+++ packages/utilities/etc/utilities.api.md
@@ -10,7 +10,6 @@ import { IStyleFunctionOrObject } from '@uifabric/merge-styles';
 import { IStyleSet } from '@uifabric/merge-styles';
 import { Omit } from '@uifabric/merge-styles';
 import * as React from 'react';
-import { ReactHTML } from 'react';
 
 // @public
 export function addDirectionalKeyCode(which: number): void;
@@ -335,7 +334,7 @@ export function getLastFocusable(rootElement: HTMLElement, currentElement: HTMLE
 export function getLastTabbable(rootElement: HTMLElement, currentElement: HTMLElement, includeElementsInFocusZones?: boolean, checkNode?: boolean): HTMLElement | null;
 
 // @public
-export function getNativeElementProps<TProps>(tagName: keyof ReactHTML, props: TProps, excludedPropNames?: string[]): TProps;
+export function getNativeElementProps<TAttributes extends React.HTMLAttributes<any>>(tagName: keyof React.ReactHTML, props: {}, excludedPropNames?: string[]): TAttributes;
 
 // @public
 export function getNativeProps<T>(props: {}, allowedPropNames: string[], excludedPropNames?: string[]): T;
